backend/fast_api/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0
pytesseract>=0.3.8
ultralytics>=8.0.0
torch>=1.7.0
scikit-image>=0.18.0
imutils>=0.5.4
pyspellchecker>=0.6.3
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0
